--- a/.changeset/olive-candles-fry.md
+++ b/.changeset/olive-candles-fry.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector-anb-no-unmatchable` performance