--- a/lib/rules/selector-anb-no-unmatchable/index.js
+++ b/lib/rules/selector-anb-no-unmatchable/index.js
@@ -46,13 +46,11 @@ const rule = (primary) => {
 		}
 
 		root.walkRules((ruleNode) => {
-			if (!isStandardSyntaxRule(ruleNode)) {
-				return;
-			}
+			if (!hasANPlusBNotationPseudoClasses(ruleNode.selector)) return;
+
+			if (!isStandardSyntaxRule(ruleNode)) return;
 
 			ruleNode.selectors.forEach((selector) => {
-				if (!hasANPlusBNotationPseudoClasses(selector)) return;
-
 				let cssTreeSelector;
 
 				try {